+++ jianfanjia/jianfanjia/Screen/RequirementList/PlanList/PlanPreviewViewController.m
@@ -58,7 +58,7 @@
     self.navigationItem.rightBarButtonItem = [[UIBarButtonItem alloc] initWithTitle:@"详细报价" style:UIBarButtonItemStylePlain target:self action:@selector(onChoosePriceDetail)];
     self.navigationItem.rightBarButtonItem.tintColor = [UIColor colorWithR:0xfe g:0x70 b:0x04];
     
-    self.title = [NSString stringWithFormat:@"方案%ld", self.order];
+    self.title = [NSString stringWithFormat:@"方案%ld", (long)self.order];
 }
 
 - (void)initUI {
